requirements.txt
streamlit 
crewai 
crewai_tools
python-dotenv 
langchain-google-genai
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
google-generativeai
crewai==0.30.11
crewai-tools==0.2.6
langchain-google-genai==1.0.3
python-dotenv
streamlit
agno
graphviz